Addresses #137.

This PR refactors the `R1CSShape` sparse matrix / dense vector multiplication API to avoid the unnecessary cloning and allocating of memory. In the codebase, we often see the pattern:
```rust
let z = [W.W.clone(), vec![U.u], U.X.clone()].concat();
let (Az, Bz, Cz) = self.multiply_vec(&z)?;
```
This is memory-inefficient. The new API directly uses references to the witness vectors, like this:
```rust
let (Az, Bz, Cz) = self.multiply_witness(&W.W, &U.u, &U.X)?;
```

Benchmark results from downstream `lurk-rs` confirm that this improves performance: https://github.com/lurk-lab/lurk-rs/pull/923

## Pasted Benchmark
  <details>
    <summary>Click to view benchmark</summary>

| Test | Base         | PR               | % |
|------|--------------|------------------|---|
| LEM Fibonacci Prove - rc = 100/fib/num-100 | 4.8±0.03s | **4.4±0.03s** | **-8.33%** |
| LEM Fibonacci Prove - rc = 100/fib/num-200 | 9.7±0.33s | **9.2±0.08s** | **-5.15%** |
| LEM Fibonacci Prove - rc = 600/fib/num-100 | 3.8±0.01s | N/A | N/A |
| LEM Fibonacci Prove - rc = 600/fib/num-200 | 8.4±0.10s | N/A | N/A |

  </details>